+++ checker/src/main/java/org/checkerframework/checker/units/UnitsAnnotationClassLoader.java
@@ -2,6 +2,7 @@ package org.checkerframework.checker.units;
 
 import java.lang.annotation.Annotation;
 import javax.lang.model.element.AnnotationMirror;
+import javax.lang.model.type.TypeMirror;
 import org.checkerframework.checker.units.qual.UnitsMultiple;
 import org.checkerframework.common.basetype.BaseTypeChecker;
 import org.checkerframework.framework.type.AnnotationClassLoader;
@@ -20,9 +21,9 @@ public class UnitsAnnotationClassLoader extends AnnotationClassLoader {
    * <p>This filter will ignore (by returning false) any units annotation which is an alias of
    * another base unit annotation (identified via {@link UnitsMultiple} meta-annotation). Alias
    * annotations can still be used in source code; they are converted into a base annotation by
-   * {@link UnitsAnnotatedTypeFactory#canonicalAnnotation(AnnotationMirror)}. This filter simply
-   * makes sure that the alias annotations themselves don't become part of the type hierarchy as
-   * their base annotations already are in the hierarchy.
+   * {@link UnitsAnnotatedTypeFactory#canonicalAnnotation(AnnotationMirror,TypeMirror)}. This filter
+   * simply makes sure that the alias annotations themselves don't become part of the type hierarchy
+   * as their base annotations already are in the hierarchy.
    */
   @Override
   protected boolean isSupportedAnnotationClass(Class<? extends Annotation> annoClass) {