+++ framework/src/main/java/org/checkerframework/common/basetype/BaseTypeVisitor.java
@@ -2191,7 +2191,7 @@ public class BaseTypeVisitor<Factory extends GenericAnnotatedTypeFactory<?, ?, ?
    * Returns true if and only if {@code inferredAnnotation} is valid for a given expression to match
    * the {@code necessaryAnnotation}.
    *
-   * <p>By default, {@code inferredAnnotation} must be a subtype of or equal to {@code
+   * <p>By default, {@code inferredAnnotation} must be a subtype of (or equal to) {@code
    * necessaryAnnotation}, but subclasses might override this behavior.
    */
   protected boolean checkContract(